+++ lib/create_exam/create_exam_form/question_card.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+class QuestionCard extends StatelessWidget {
+  const QuestionCard({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          const ListTile(
+            leading: Text('1'),
+            title: Text('Leg het verschil uit tussen een stack en een heap.'),
+            subtitle: Text(
+              'Open vraag',
+              style: TextStyle(fontFamily: 'cursive'),
+            ),
+            trailing: Text('2 ptn.'),
+          ),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: <Widget>[
+              IconButton(
+                icon: const Icon(Icons.delete),
+                onPressed: () => print('delete'),
+              ),
+              const SizedBox(width: 8),
+              IconButton(
+                icon: const Icon(Icons.edit),
+                onPressed: () => print('edit'),
+              ),
+              const SizedBox(width: 8),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
